♻️ refactor: 프로젝트 등록 스토어를 Context 스코프로 전환
전역 싱글톤 Zustand 스토어를 createStore 팩토리 + React Context로 전환한다. 등록 페이지 마운트 시 스토어 인스턴스를 생성하고 언마운트 시 GC로 소멸시켜, 중도 이탈/재진입/재등록 시 이전 폼 데이터가 남던 상태 오염을 구조적으로 해결한다. 하이드레이터는 스토어 인스턴스를 인자로 받도록 변경했고, 소비 컴포넌트는 셀렉터 시그니처가 동일해 무변경.

#562
