+++ src/features/project/new/ui/ProjectRegisterPage.tsx
@@ -38,7 +38,11 @@ import { hydrateApplicationFormIntoStore } from "@/features/project/new/model/ap
 import { hydrateDraftIntoStore } from "@/features/project/new/model/draftHydrator"
 import { isWithinMatchingPeriod } from "@/features/project/new/model/matchingPeriod"
 import { hydrateProjectDetailIntoStore } from "@/features/project/new/model/projectDetailHydrator"
-import { useProjectRegisterStore } from "@/features/project/new/model/useProjectRegisterStore"
+import {
+  ProjectRegisterStoreProvider,
+  useProjectRegisterStore,
+  useProjectRegisterStoreApi,
+} from "@/features/project/new/model/useProjectRegisterStore"
 import { useActiveGisu } from "@/shared/hooks/useActiveGisu"
 import { CtaModal } from "@/shared/ui/modal/CtaModal"
 import { useToastStore } from "@/shared/ui/toast/useToastStore"
@@ -49,7 +53,7 @@ type ProjectRegisterPageProps =
   | { mode: "new" }
   | { mode: "edit"; editProjectId: number }
 
-export function ProjectRegisterPage(props: ProjectRegisterPageProps) {
+function ProjectRegisterPageContent(props: ProjectRegisterPageProps) {
   const isEditMode = props.mode === "edit"
   const editProjectId = props.mode === "edit" ? props.editProjectId : undefined
   const [step, setStep] = useState(1)
@@ -83,6 +87,7 @@ export function ProjectRegisterPage(props: ProjectRegisterPageProps) {
     if (!matchingRoundsQuery.isSuccess) return true
     return isWithinMatchingPeriod(matchingRoundsQuery.data, new Date())
   }, [isEditMode, matchingRoundsQuery.isSuccess, matchingRoundsQuery.data])
+  const storeApi = useProjectRegisterStoreApi()
   const projectId = useProjectRegisterStore((s) => s.projectId)
   const application = useProjectRegisterStore((s) => s.application)
   const pmInfo = useProjectRegisterStore((s) => s.pmInfo)
@@ -116,7 +121,7 @@ export function ProjectRegisterPage(props: ProjectRegisterPageProps) {
   const resolveCanEditRecruitStep = async (): Promise<boolean> => {
     if (!isEditMode) return false
 
-    const pid = useProjectRegisterStore.getState().projectId
+    const pid = storeApi.getState().projectId
     if (pid === null) return false
     try {
       const permission = await queryClient.ensureQueryData({
@@ -225,9 +230,9 @@ export function ProjectRegisterPage(props: ProjectRegisterPageProps) {
 
   useEffect(() => {
     if (detailQuery.data) {
-      hydrateProjectDetailIntoStore(detailQuery.data)
+      hydrateProjectDetailIntoStore(detailQuery.data, storeApi)
     }
-  }, [detailQuery.data])
+  }, [detailQuery.data, storeApi])
 
   useEffect(() => {
     if (!isEditMode || !detailQuery.isError) return
@@ -273,13 +278,13 @@ export function ProjectRegisterPage(props: ProjectRegisterPageProps) {
   useEffect(() => {
     if (isEditMode) return
     if (draftQuery.data && !projectId) {
-      hydrateDraftIntoStore(draftQuery.data)
+      hydrateDraftIntoStore(draftQuery.data, storeApi)
     }
-  }, [isEditMode, draftQuery.data, projectId])
+  }, [isEditMode, draftQuery.data, projectId, storeApi])
 
   useEffect(() => {
     return () => {
-      const { gisuId: storedGisuId } = useProjectRegisterStore.getState()
+      const { gisuId: storedGisuId } = storeApi.getState()
 
       reset()
       if (storedGisuId) {
@@ -288,7 +293,7 @@ export function ProjectRegisterPage(props: ProjectRegisterPageProps) {
         })
       }
     }
-  }, [queryClient, reset])
+  }, [queryClient, reset, storeApi])
 
   const applicationFormQuery = useQuery({
     queryKey: projectKeys.applicationForm(projectId ?? 0),
@@ -299,7 +304,7 @@ export function ProjectRegisterPage(props: ProjectRegisterPageProps) {
   useEffect(() => {
     if (!isEditMode) return
     if (applicationFormQuery.data) {
-      hydrateApplicationFormIntoStore(applicationFormQuery.data)
+      hydrateApplicationFormIntoStore(applicationFormQuery.data, storeApi)
       setApplicationFormHydrated(true)
     } else if (applicationFormQuery.data === null) {
       setApplication({
@@ -309,7 +314,7 @@ export function ProjectRegisterPage(props: ProjectRegisterPageProps) {
       })
       setApplicationFormHydrated(true)
     }
-  }, [isEditMode, applicationFormQuery.data, setApplication])
+  }, [isEditMode, applicationFormQuery.data, setApplication, storeApi])
 
   const submitMutation = useMutation({
     mutationFn: async () => {
@@ -638,5 +643,13 @@ export function ProjectRegisterPage(props: ProjectRegisterPageProps) {
         onConfirm={() => void handleSaveAndLeave()}
       />
     </section>
+  )
+}
+
+export function ProjectRegisterPage(props: ProjectRegisterPageProps) {
+  return (
+    <ProjectRegisterStoreProvider>
+      <ProjectRegisterPageContent {...props} />
+    </ProjectRegisterStoreProvider>
   )
 }
